requirements.txt
torch==2.2.0
torchvision==0.17.0
pandas==2.2.0
numpy==1.26.4
Pillow==10.2.0
scikit-learn==1.4.0
streamlit==1.32.0